+++ lib/src/theme/macos_theme.dart
@@ -189,7 +189,7 @@ class MacosThemeData with Diagnosticable {
     Brightness? brightness,
     Color? primaryColor,
     Color? canvasColor,
-    Typography? typography,
+    MacosTypography? typography,
     PushButtonThemeData? pushButtonTheme,
     Color? dividerColor,
     HelpButtonThemeData? helpButtonTheme,
@@ -205,7 +205,7 @@ class MacosThemeData with Diagnosticable {
     canvasColor ??= isDark
         ? CupertinoColors.systemBackground.darkElevatedColor
         : CupertinoColors.systemBackground;
-    typography ??= Typography(
+    typography ??= MacosTypography(
       color: brightness == Brightness.light
           ? CupertinoColors.black
           : CupertinoColors.white,
@@ -299,7 +299,7 @@ class MacosThemeData with Diagnosticable {
   final Color canvasColor;
 
   /// The default text styling for this theme.
-  final Typography typography;
+  final MacosTypography typography;
 
   /// The default style for [PushButton]s below the overall [MacosTheme].
   final PushButtonThemeData pushButtonTheme;
@@ -311,7 +311,7 @@ class MacosThemeData with Diagnosticable {
   /// The default style for [HelpButton]s below the overall [MacosTheme].
   final HelpButtonThemeData helpButtonTheme;
 
-  /// The default style for [Tooltip]s below the overall [MacosTheme]
+  /// The default style for [MacosTooltip]s below the overall [MacosTheme]
   final TooltipThemeData tooltipTheme;
 
   /// The density value for specifying the compactness of various UI components.
@@ -319,7 +319,7 @@ class MacosThemeData with Diagnosticable {
   /// {@macro flutter.material.themedata.visualDensity}
   final VisualDensity visualDensity;
 
-  /// The default style for [Scrollbar]s below the overall [MacosTheme]
+  /// The default style for [MacosScrollbar]s below the overall [MacosTheme]
   final ScrollbarThemeData scrollbarTheme;
 
   /// Linearly interpolate between two themes.
@@ -329,7 +329,7 @@ class MacosThemeData with Diagnosticable {
       dividerColor: Color.lerp(a.dividerColor, b.dividerColor, t)!,
       primaryColor: Color.lerp(a.primaryColor, b.primaryColor, t)!,
       canvasColor: Color.lerp(a.primaryColor, b.primaryColor, t)!,
-      typography: Typography.lerp(a.typography, b.typography, t),
+      typography: MacosTypography.lerp(a.typography, b.typography, t),
       helpButtonTheme:
           HelpButtonThemeData.lerp(a.helpButtonTheme, b.helpButtonTheme, t),
       pushButtonTheme:
@@ -346,7 +346,7 @@ class MacosThemeData with Diagnosticable {
     Brightness? brightness,
     Color? primaryColor,
     Color? canvasColor,
-    Typography? typography,
+    MacosTypography? typography,
     PushButtonThemeData? pushButtonTheme,
     Color? dividerColor,
     HelpButtonThemeData? helpButtonTheme,
@@ -375,7 +375,8 @@ class MacosThemeData with Diagnosticable {
     properties.add(ColorProperty('primaryColor', primaryColor));
     properties.add(ColorProperty('canvasColor', canvasColor));
     properties.add(ColorProperty('dividerColor', dividerColor));
-    properties.add(DiagnosticsProperty<Typography>('typography', typography));
+    properties
+        .add(DiagnosticsProperty<MacosTypography>('typography', typography));
     properties.add(DiagnosticsProperty<PushButtonThemeData>(
       'pushButtonTheme',
       pushButtonTheme,